+++ packages/query/src/block-processor.ts
@@ -310,6 +310,7 @@ export class BlockProcessor implements BlockProcessorInterface {
         spentNullifiers,
         recordsByCommitment,
         blockTx,
+        addr => this.viewServer.isControlledAddress(addr),
       );
 
       // this simply stores the new records with 'rehydrated' sources to idb