+++ packages/query/src/helpers/identify-txs.test.ts
@@ -1,8 +1,5 @@
 import { describe, expect, test } from 'vitest';
-import {
-  CommitmentSource,
-  Nullifier,
-} from '@penumbra-zone/protobuf/penumbra/core/component/sct/v1/sct_pb';
+import { Nullifier } from '@penumbra-zone/protobuf/penumbra/core/component/sct/v1/sct_pb';
 import { StateCommitment } from '@penumbra-zone/protobuf/penumbra/crypto/tct/v1/tct_pb';
 import {
   Action,
@@ -10,6 +7,7 @@ import {
   TransactionBody,
 } from '@penumbra-zone/protobuf/penumbra/core/transaction/v1/transaction_pb';
 import {
+  BLANK_TX_SOURCE,
   getCommitmentsFromActions,
   getNullifiersFromActions,
   identifyTransactions,
@@ -27,10 +25,15 @@ import {
   SwapClaimBody,
 } from '@penumbra-zone/protobuf/penumbra/core/component/dex/v1/dex_pb';
 import { SpendableNoteRecord, SwapRecord } from '@penumbra-zone/protobuf/penumbra/view/v1/view_pb';
-
-const BLANK_TX_SOURCE = new CommitmentSource({
-  source: { case: 'transaction', value: { id: new Uint8Array() } },
-});
+import { Any } from '@bufbuild/protobuf';
+import {
+  FungibleTokenPacketData,
+  IbcRelay,
+} from '@penumbra-zone/protobuf/penumbra/core/component/ibc/v1/ibc_pb';
+import { addressFromBech32m } from '@penumbra-zone/bech32m/penumbra';
+import { Address } from '@penumbra-zone/protobuf/penumbra/core/keys/v1/keys_pb';
+import { Packet } from '@penumbra-zone/protobuf/ibc/core/channel/v1/channel_pb';
+import { MsgRecvPacket } from '@penumbra-zone/protobuf/ibc/core/channel/v1/tx_pb';
 
 describe('getCommitmentsFromActions', () => {
   test('returns empty array when tx.body.actions is undefined', () => {
@@ -213,7 +216,12 @@ describe('identifyTransactions', () => {
     const spentNullifiers = new Set<Nullifier>();
     const commitmentRecords = new Map<StateCommitment, SpendableNoteRecord | SwapRecord>();
 
-    const result = await identifyTransactions(spentNullifiers, commitmentRecords, blockTx);
+    const result = await identifyTransactions(
+      spentNullifiers,
+      commitmentRecords,
+      blockTx,
+      () => false,
+    );
 
     expect(result.relevantTxs).toEqual([]);
     expect(result.recoveredSourceRecords).toEqual([]);
@@ -311,12 +319,17 @@ describe('identifyTransactions', () => {
 
     const spentNullifiersBeforeSize = spentNullifiers.size;
     const commitmentRecordsBeforeSize = commitmentRecords.size;
-    const result = await identifyTransactions(spentNullifiers, commitmentRecords, [
-      tx1, // relevant
-      tx2, // relevant
-      tx3, // not
-      tx4, // not
-    ]);
+    const result = await identifyTransactions(
+      spentNullifiers,
+      commitmentRecords,
+      [
+        tx1, // relevant
+        tx2, // relevant
+        tx3, // not
+        tx4, // not
+      ],
+      () => false,
+    );
 
     expect(result.relevantTxs.length).toBe(2);
     expect(result.recoveredSourceRecords.length).toBe(1);
@@ -328,4 +341,40 @@ describe('identifyTransactions', () => {
     expect(spentNullifiersBeforeSize).toEqual(spentNullifiers.size);
     expect(commitmentRecordsBeforeSize).toEqual(commitmentRecords.size);
   });
+
+  test('identifies ibc relays', async () => {
+    const knownAddr =
+      'penumbra1e8k5cyds484dxvapeamwveh5khqv4jsvyvaf5wwxaaccgfghm229qw03pcar3ryy8smptevstycch0qk3uu0rgkvtjpxy3cu3rjd0agawqtlz6erev28a6sg69u7cxy0t02nd4';
+    const unknownAddr =
+      'penumbracompat1147mfall0zr6am5r45qkwht7xqqrdsp50czde7empv7yq2nk3z8yyfh9k9520ddgswkmzar22vhz9dwtuem7uxw0qytfpv7lk3q9dp8ccaw2fn5c838rfackazmgf3ahhwqq0da';
+    const tx = new Transaction({
+      body: {
+        actions: [createIbcRelay(knownAddr), createIbcRelay(unknownAddr)],
+      },
+    });
+    const blockTx = [tx];
+    const spentNullifiers = new Set<Nullifier>();
+    const commitmentRecords = new Map<StateCommitment, SpendableNoteRecord | SwapRecord>();
+
+    const result = await identifyTransactions(spentNullifiers, commitmentRecords, blockTx, addr =>
+      addr.equals(new Address(addressFromBech32m(knownAddr))),
+    );
+
+    expect(result.relevantTxs.length).toBe(1);
+    expect(result.relevantTxs[0]?.data.equals(tx)).toBeTruthy();
+    expect(result.recoveredSourceRecords.length).toBe(0);
+  });
 });
+
+const createIbcRelay = (receiver: string): Action => {
+  const tokenPacketData = new FungibleTokenPacketData({ receiver });
+  const encoder = new TextEncoder();
+  const relevantRelay = Any.pack(
+    new MsgRecvPacket({
+      packet: new Packet({ data: encoder.encode(tokenPacketData.toJsonString()) }),
+    }),
+  );
+  return new Action({
+    action: { case: 'ibcRelayAction', value: new IbcRelay({ rawAction: relevantRelay }) },
+  });
+};