+++ packages/storage/src/indexed-db/index.ts
@@ -64,6 +64,7 @@ import { PartialMessage, PlainMessage } from '@bufbuild/protobuf';
 import { getAmountFromRecord } from '@penumbra-zone/getters/spendable-note-record';
 import { isZero } from '@penumbra-zone/types/amount';
 import { IDB_VERSION } from './config.js';
+import { typeRegistry } from '@penumbra-zone/protobuf';
 
 const assertBytes = (v?: Uint8Array, expect?: number, name = 'value'): v is Uint8Array => {
   if (expect !== undefined && v?.length !== expect) {
@@ -363,7 +364,7 @@ export class IndexedDb implements IndexedDbInterface {
     const tx = new TransactionInfo({ id, height, transaction });
     await this.u.update({
       table: 'TRANSACTIONS',
-      value: tx.toJson() as Jsonified<TransactionInfo>,
+      value: tx.toJson({ typeRegistry }) as Jsonified<TransactionInfo>,
     });
   }
 
@@ -374,7 +375,7 @@ export class IndexedDb implements IndexedDbInterface {
     if (!jsonRecord) {
       return undefined;
     }
-    return TransactionInfo.fromJson(jsonRecord);
+    return TransactionInfo.fromJson(jsonRecord, { typeRegistry });
   }
 
   async getFmdParams(): Promise<FmdParameters | undefined> {