+++ packages/ui/src/CharacterTransition/index.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { CharacterTransition } from '.';
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 
 const WhiteTextWrapper = styled.div`
   color: ${props => props.theme.color.text.primary};