+++ packages/ui/src/Slider/index.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import styled, { css } from 'styled-components';
+import { styled, css } from 'styled-components';
 import * as RadixSlider from '@radix-ui/react-slider';
 import { detail, detailTechnical } from '../utils/typography';
 import { theme } from '../PenumbraUIProvider/theme';