+++ apps/node-status/CHANGELOG.md
@@ -1,5 +1,26 @@
 # node-status
 
+## 4.1.28
+
+### Patch Changes
+
+- Updated dependencies [b6e32f8]
+- Updated dependencies [b6e32f8]
+  - @penumbra-zone/protobuf@6.1.0
+  - @penumbra-zone/types@24.0.0
+  - @penumbra-zone/ui@10.0.0
+  - @penumbra-zone/crypto-web@25.0.0
+
+## 4.1.27
+
+### Patch Changes
+
+- Updated dependencies [b1d4b7d]
+- Updated dependencies [75ef4f5]
+- Updated dependencies [de9bd06]
+- Updated dependencies [43d8350]
+  - @penumbra-zone/ui@9.2.0
+
 ## 4.1.26
 
 ### Patch Changes